Improve extension endpoint handling, add apis, and add extension event handlers
This moves the endpoint handling to the FavaExtensionBase class, and
adds the extension_endpoint decorator to allow extensions to register
endpoints more easily.

It also adds the ExtensionApi class which is set on the extension
module to allow the extension to call it's endpoints more easily.

Finally, it adds the initExtensionHandlers which looks for elements
with the class "extension-handler" and automatically converts all of
their "data-handler-<event>" attributes to event handlers that can
access the extension js module as "this". This allows templates to
fill in a part of the handler programmatically, which is useful when
e.g. a number of buttons are created in a for loop in the template,
and each button should do something based on the loop value.