Improve extension endpoint handling, and add apis
This moves the endpoint handling to the FavaExtensionBase class, and
adds the extension_endpoint decorator to allow extensions to register
endpoints more easily.

It also adds the ExtensionContext and class which is passed to all
extension module functions. The context currently only contains an
ExtensionApi, which allows the extension to call it's endpoints more
easily.